Fixes and updates

+++ install.sh
@@ -19,23 +19,29 @@ for f in .zshrc .venv; do
     # Already backed up
     rm "${HOME}/${f}"
   else
-    # Backup
-    mv "${HOME}/${f}" "${me}/backup/${f}"
+    if [[ -e "${HOME}/${f}" ]]; then
+        # Backup
+        mv "${HOME}/${f}" "${me}/backup/${f}"
+    fi
   fi
   ln -s "${me}/${f}" "${HOME}/${f}"
 done
 
 git config --global include.path "${me}/.gitconfig"
 
-# Default virtualenv
-export WORKON_HOME="${HOME}/VirtualEnv"
-
-
 # Default tree
+mkdir -p "${HOME}/virtualenv"
 mkdir -p "${HOME}/Projects"
 mkdir -p "${HOME}/Applications"
 
+# Default virtualenv
+export WORKON_HOME="${HOME}/virtualenv"
+
+# Switch to zsh
 chsh -s /bin/zsh
+
 echo "You should re-login to make available for all sessions"
 echo "Remember to select a Powerline font for the terminal"
 echo "To install extra goodies run ${me}/bin/install_goodies.sh"
+echo "To install Gnome extensions run ${me}/bin/install_gnome_extensions.sh"
+
